Fix attribute assignment
The attribute fields in the Java classes were org.python.Objects but when trying to
assign a new value to a field, the new value would be converted to a Java value.
'date.year = 12' would cast the 12 from a org.python.types.Int to a Java long, although the year attribute is a org.python.Object.