Prevent adding of new attributes to Date
In python, classes can specify __slots__ which, among other things, can prevent them having a __dict__ field.
In that case, no attributes can be added to the class.
This __slots__ mechanism is not implemented in VOC, thus this fix is needed to keep the behavior consistent with the python implementation.